need to implement password & i added one more file to channel
