+++ src/Channel/ChannelManagement.cpp
@@ -3,7 +3,6 @@
 #include <stdexcept>
 #include <algorithm>
 
-// Client management
 const std::vector<Client*> &Channel::getClients() const
 {
     return (_clients);
@@ -30,36 +29,6 @@ bool Channel::hasClient(Client *client) const
 {
     auto it = std::find(_clients.begin(), _clients.end(), client);
     if (it == _clients.end())
-        return (false);
-    return (true);
-}
-
-// Operator management
-const std::set<Client*> &Channel::getOperators() const
-{
-    return (_operators);
-}
-
-void Channel::addOperator(Client *client)
-{
-    _operators.insert(client);
-    // We can use the return value if the client is allready
-    // an operator. Should we ?
-}
-
-void Channel::removeOperator(Client *client)
-{
-    auto it = getOperators().find(client);
-    if (it == getOperators().end());
-        throw std::runtime_error("Client is not in the invitation list.");
-    // Maybe we need to implement a different message error
-    _operators.erase(client);
-}
-
-bool Channel::isOperator(Client *client) const
-{
-    auto it = getOperators().find(client);
-    if (it == getOperators().end())
         return (false);
     return (true);
 }
